+++ classes/hydraretrieve.py
@@ -7,22 +7,25 @@
 
 
 # Import libraries
+from urllib import request
 
 # Import script modules
 
 
-# Retrieve data via beacon files or APIs
 class HydraRetrieve:
 
     something = None
 
 
-    def __init__(self, something:str = ''):
+    def __init__(self, command, output, report, morph):
         '''
-        Add required data to instances of this object
+        Retrieve data via beacon files or APIs
 
             Parameters:
-                something (str): ???
+                command (str): ???
+                output (str): ???
+                report (str): ???
+                morph (str): ???
         '''
 
         # Assign variables
@@ -36,82 +39,6 @@ class HydraRetrieve:
 
         # Put together a string
         return self.something
-
-# types: api, resources
-
-# # Set up Hydra API routine if required
-# if 'lists' in command.request or 'beacon' in command.request or 'list_triples' in command.request or 'list_cgif' in command.request:
-#     hydra = Hydra(command.target_folder_name, command.source_url, command.content_type)
-
-#     # Populate the object, and download each API list if requested
-#     if 'lists' in command.request:
-#         save_lists = True
-#     else:
-#         save_lists = False
-#     hydra.populate(save_lists, command.resource_url_filter, command.resource_url_replace, command.resource_url_replace_with, command.resource_url_add)
-
-#     # Compile a beacon list if requested
-#     if 'beacon' in command.request:
-#         hydra.save_beacon()
-
-#     # Compile list triples if requested
-#     if 'list_triples' in command.request:
-#         hydra.save_triples()
-
-#     # Compile list CGIF triples if requested
-#     if 'list_cgif' in command.request:
-#         hydra.save_triples('cgif', 'lists_cgif')
-
-#     # Add status message
-#     status.extend(hydra.status)
-
-# # Mini Hydra routine if Beacon logic is requested but no beacon file is given
-# elif command.source_file == '':
-#     hydra = Hydra(command.target_folder_name, command.source_url, command.content_type)
-#     hydra.populate(False, command.resource_url_filter, command.resource_url_replace, command.resource_url_replace_with, command.resource_url_add)
-
-# # Mark absence of hydra object if beacon file is present
-# else:
-#     hydra = False
-
-# # Set up beacon file routine if required
-# if 'resources' in command.request or 'resource_triples' in command.request or 'resource_cgif' in command.request or 'resource_table' in command.request:
-
-#     # Use previous resource list if present
-#     if hydra == False:
-#         beacon = Beacon(command.target_folder_name, command.content_type)
-#     else:
-#         beacon = Beacon(command.target_folder_name, command.content_type, hydra.resources)
-
-#     # Populate the object, and download each resource if requested
-#     if 'resources' in command.request:
-#         save_resources = True
-#     else:
-#         save_resources = False
-#     beacon.populate(save_resources, command.clean_resource_names, command.source_file, command.source_folder, command.supplement_data_feed, command.supplement_data_catalog, command.supplement_data_catalog_publisher)
-
-#     # Compile resource triples if requested
-#     if 'resource_triples' in command.request:
-#         beacon.save_triples()
-
-#     # Compile resource CGIF triples if requested
-#     if 'resource_cgif' in command.request:
-#         beacon.save_triples('cgif', 'resources_cgif')
-
-#     # Compile resource table if requested
-#     if 'resource_table' in command.request:
-#         beacon.save_csv(command.table_data)
-
-#     # Add status message
-#     status.extend(beacon.status)
-
-
-
-# # Import libraries
-# from urllib import request
-
-# # Import script modules
-# from helpers.clean import _strip_lines
 
 
 # def download_file(url:str, content_type:str = '') -> dict:
@@ -284,6 +211,83 @@ class HydraRetrieve:
 #         return file_extension
 #     else:
 #         return file_type
+
+
+
+
+
+
+
+# types: api, resources
+
+# # Set up Hydra API routine if required
+# if 'lists' in command.download or 'beacon' in command.download or 'list_triples' in command.download or 'list_cgif' in command.download:
+#     hydra = Hydra(command.target_folder_path, command.source_url, command.content_type)
+
+#     # Populate the object, and download each API list if requested
+#     if 'lists' in command.download:
+#         save_lists = True
+#     else:
+#         save_lists = False
+#     hydra.populate(save_lists, command.resource_url_filter, command.resource_url_replace, command.resource_url_replace_with, command.resource_url_add)
+
+#     # Compile a beacon list if requested
+#     if 'beacon' in command.download:
+#         hydra.save_beacon()
+
+#     # Compile list triples if requested
+#     if 'list_triples' in command.download:
+#         hydra.save_triples()
+
+#     # Compile list CGIF triples if requested
+#     if 'list_cgif' in command.download:
+#         hydra.save_triples('cgif', 'lists_cgif')
+
+#     # Add status message
+#     status.extend(hydra.status)
+
+# # Mini Hydra routine if Beacon logic is requested but no beacon file is given
+# elif command.source_file == '':
+#     hydra = Hydra(command.target_folder_path, command.source_url, command.content_type)
+#     hydra.populate(False, command.resource_url_filter, command.resource_url_replace, command.resource_url_replace_with, command.resource_url_add)
+
+# # Mark absence of hydra object if beacon file is present
+# else:
+#     hydra = False
+
+# # Set up beacon file routine if required
+# if 'resources' in command.download or 'resource_triples' in command.download or 'resource_cgif' in command.download or 'resource_table' in command.download:
+
+#     # Use previous resource list if present
+#     if hydra == False:
+#         beacon = Beacon(command.target_folder_path, command.content_type)
+#     else:
+#         beacon = Beacon(command.target_folder_path, command.content_type, hydra.resources)
+
+#     # Populate the object, and download each resource if requested
+#     if 'resources' in command.download:
+#         save_resources = True
+#     else:
+#         save_resources = False
+#     beacon.populate(save_resources, command.clean_resource_names, command.source_file, command.source_folder, command.supplement_data_feed, command.supplement_data_catalog, command.supplement_data_catalog_publisher)
+
+#     # Compile resource triples if requested
+#     if 'resource_triples' in command.download:
+#         beacon.save_triples()
+
+#     # Compile resource CGIF triples if requested
+#     if 'resource_cgif' in command.download:
+#         beacon.save_triples('cgif', 'resources_cgif')
+
+#     # Compile resource table if requested
+#     if 'resource_table' in command.download:
+#         beacon.save_csv(command.table_data)
+
+#     # Add status message
+#     status.extend(beacon.status)
+
+
+
 
 
 #------------------------------------------------------------------